fix: keep remove twins functions to naked twins and hidden twins, independently.
The initial idea was to implement find_hidden_twins() and find_hidden_twins() and use the same function _remove_twins() to remove the matching cells. Due to the lack of time, it has been applied the same approach used on the naked twins to the hidden twins pairs, that is to say find_, remove_, and strategy(naked or hidden) main call.